AI/requirement.txt
absl-py==2.1.0
aiohttp==3.9.5
astunparse==1.6.3
attrs==23.2.0
bottleneck==1.3.7
certifi==2024.6.2
click==8.1.7
datasets==2.19.2
debugpy==1.6.7
dill==0.3.8
fastapi==0.115.0
filelock==3.15.4
flask==2.2.2
flatbuffers==24.3.25
gast==0.5.4
grpcio==1.64.1
h5py==3.11.0
httpx==0.27.0
huggingface_hub==0.23.4
ipykernel==6.29.4
ipython==8.25.0
joblib==1.4.2
keras==3.3.3
langchain==0.2.8
matplotlib==3.6.3
numpy==1.23.4
opencv-python==4.8.0.74
pandas==1.5.3
pillow==9.4.0
pyarrow==13.0.0
pyparsing==3.0.9
scikit-learn==1.2.2
scipy==1.10.1
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.27.0
uvicorn==0.23.2
